Added Graph and added to Node
Graph creates a 2D array of Nodes

Node's removeNode() works now, using a loop to iterate through the
vector until the correct memory address is found of the Node to be
removed.
Node's addNode() now takes in a pointer to the Node to be added.
Node* must be used instead of Node in the vector to prevent errors with
scope.